+++ src/@lib/context/notification/index.ts
@@ -1,0 +1,16 @@
+import { createContext, useContext } from "react";
+import { NotificationContextType } from "../../../types";
+
+export const NotificationContext = createContext<
+  NotificationContextType | undefined
+>(undefined);
+
+export const useNotification = () => {
+  const context = useContext(NotificationContext);
+  if (context === undefined) {
+    throw new Error(
+      "useNotification must be used within a NotificationProvider",
+    );
+  }
+  return context;
+};
