+++ src/@lib/context/theme/index.ts
@@ -1,0 +1,14 @@
+import { createContext, useContext } from "react";
+import { ThemeContextType } from "../../../types";
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(
+  undefined,
+);
+
+export const useTheme = () => {
+  const context = useContext(ThemeContext);
+  if (context === undefined) {
+    throw new Error("useTheme must be used within an AppProvider");
+  }
+  return context;
+};
