--- src/@lib/hooks/useProviders.tsx
@@ -1,91 +1,0 @@
-import { PropsWithChildren, useState } from "react";
-import { Notification, User } from "../../types";
-import { useCallback } from "./useCallback";
-import {
-  AppContext,
-  NotificationContext,
-  ThemeContext,
-  useNotification,
-} from "./useContext";
-import { useMemo } from "./useMemo";
-
-export const AppProvider: React.FC<PropsWithChildren> = ({ children }) => {
-  const [user, setUser] = useState<User | null>(null);
-  const { addNotification } = useNotification();
-
-  const login = useCallback((email: string) => {
-    setUser({ id: 1, name: "홍길동", email });
-    addNotification("로그인 되었습니다.", "success");
-  }, []);
-
-  const logout = useCallback(() => {
-    setUser(null);
-    addNotification("로그아웃 되었습니다.", "success");
-  }, []);
-
-  const value = useMemo(
-    () => ({
-      user,
-      login,
-      logout,
-    }),
-    [user, login, logout],
-  );
-
-  return <AppContext.Provider value={value}>{children}</AppContext.Provider>;
-};
-
-export const ThemeProvider: React.FC<PropsWithChildren> = ({ children }) => {
-  const [theme, setTheme] = useState<"light" | "dark">("light");
-
-  const toggleTheme = useCallback(() => {
-    setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
-  }, []);
-
-  const value = useMemo(() => ({ theme, toggleTheme }), [theme, toggleTheme]);
-  return (
-    <ThemeContext.Provider value={value}>{children}</ThemeContext.Provider>
-  );
-};
-
-export const NotificationProvider: React.FC<PropsWithChildren> = ({
-  children,
-}) => {
-  const [notifications, setNotifications] = useState<Notification[]>([]);
-
-  const addNotification = useCallback(
-    (message: string, type: Notification["type"]) => {
-      const newNotification: Notification = {
-        id: Date.now(),
-        message,
-        type,
-      };
-      setNotifications((prev) => [...prev, newNotification]);
-    },
-    [setNotifications],
-  );
-
-  const removeNotification = useCallback(
-    (id: number) => {
-      setNotifications((prev) =>
-        prev.filter((notification) => notification.id !== id),
-      );
-    },
-    [setNotifications],
-  );
-
-  const value = useMemo(
-    () => ({
-      notifications,
-      addNotification,
-      removeNotification,
-    }),
-    [notifications, addNotification, removeNotification],
-  );
-
-  return (
-    <NotificationContext.Provider value={value}>
-      {children}
-    </NotificationContext.Provider>
-  );
-};
